backend/requirements.txt
Django==5.2.4
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0
django-environ==0.11.2
djangorestframework==3.14.0
dj-database-url==3.0.1
#new 9/07/25
djangorestframework-simplejwt==5.3.1
